backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
copilotkit==0.1.26
langgraph==0.4.1
langchain-groq==0.3.5
langchain-core==1.2.22
python-dotenv==1.0.1
python-multipart==0.0.22
groq==0.9.0
